:lipstick: Move UUIDBytes class to Utils
